+++ embassy-usb-synopsys-otg/CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- next-header -->
+
 ## Unreleased - ReleaseDate
 
 - Fixed: Periodic IN transfer future can no longer park forever.
 - Fixed: Later events no longer override earlier ones.
+- Fixed: Correctly work with low-speed devices over high-speed hubs.
 
 ## 0.4.0 - 2026-05-28